move perm importance and define per col method
